requirements.txt
RPi.GPIO==0.7.0
mfrc522==0.0.7
Adafruit-DHT==1.4.0
RPLCD==1.3.1
paho-mqtt==1.6.1
python-decouple